rename <Gas/> to <TXMetaDataPanel/>
