--- common/components/Gas/index.tsx
@@ -1,2 +1,0 @@
-import Gas from './Gas';
-export default Gas;
