+++ common/components/TXMetaDataPanel/TXMetaDataPanel.scss
