+++ common/components/TXMetaDataPanel/components/SimpleGas.scss
