+++ common/components/TXMetaDataPanel/index.tsx
@@ -1,0 +1,2 @@
+import TXMetaDataPanel from './TXMetaDataPanel';
+export default TXMetaDataPanel;
