+++ common/containers/Tabs/Swap/components/LiteSend/Fields.tsx
@@ -3,7 +3,7 @@ import { AmountFieldFactory } from 'components/AmountFieldFactory';
 import { AddressFieldFactory } from 'components/AddressFieldFactory';
 import { connect } from 'react-redux';
 import { AppState } from 'reducers';
-import { GenerateTransaction, SendButton, SigningStatus, Gas } from 'components';
+import { GenerateTransaction, SendButton, SigningStatus, TXMetaDataPanel } from 'components';
 import { resetWallet, TResetWallet } from 'actions/wallet';
 import translate from 'translations';
 import { getUnit } from 'selectors/transaction';
@@ -76,7 +76,7 @@ class FieldsClass extends Component<Props> {
         </div>
         <div className="row form-group">
           <div className="col-xs-12">
-            <Gas initialState={'simple'} disableToggle={true} />
+            <TXMetaDataPanel initialState={'simple'} disableToggle={true} />
           </div>
         </div>
         <SigningStatus />
